Revise `BMPDraw.cpp` as a template function and add to Adafruit_GFX library
* Allows use with a wide range of TFT displays without requiring a virtual base class
* Eliminate code duplication
