+++ samples/ScreenTFT_ILI9340-ILI9341/app/application.cpp
@@ -1,7 +1,6 @@
 #include <SmingCore.h>
 #include <Libraries/Adafruit_ILI9341/Adafruit_ILI9341.h>
-
-#include "BPMDraw.h"
+#include <Libraries/Adafruit_GFX/BMPDraw.h>
 
 // If you want, you can define WiFi settings globally in Eclipse Environment Variables
 #ifndef WIFI_SSID
