+++ samples/ScreenTFT_ST7735/app/application.cpp
@@ -1,7 +1,6 @@
 #include <SmingCore.h>
 #include <Libraries/Adafruit_ST7735/Adafruit_ST7735.h>
-
-#include "BPMDraw.h"
+#include <Libraries/Adafruit_GFX/BMPDraw.h>
 
 /*
  * Hardware SPI mode:
